requirements.txt
Pillow
openai
requests